+++ src/pages/IssueList/issuelist.tsx
@@ -1,11 +1,10 @@
-
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import Grid from '@material-ui/core/Grid'
+import Grid from '@material-ui/core/Grid';
 import Container from '@material-ui/core/Container';
-import AppbarShift from './appbarshift';
-import IssueTable from './issuetable';
+import AppbarShift from '../layout/appbarshift';
+import IssueTable from './components/issuetable';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -23,7 +22,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function Issue():JSX.Element {
+export default function Issue(): JSX.Element {
   const classes = useStyles();
   return (
     <div className={classes.root}>
